FlaskGuiTemplate/requirements.txt
bidict==0.21.2
click==7.1.2
dnspython==1.16.0
eventlet==0.30.2
Flask==1.1.2
Flask-SocketIO==5.0.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-engineio==4.1.0
python-socketio==5.2.1
six==1.15.0
SQLAlchemy==1.4.11
Werkzeug==1.0.1
